Change repository name and cretae error foulder
