+++ blockchain-watcher/src/domain/errors/index.ts
@@ -1,0 +1,2 @@
+export * from "./Solana";
+export * from "./Fallible";
